fuck you heroku

+++ app.py
@@ -29,8 +29,8 @@ class Schematic(db.Model):
 class Point(db.Model):
     __tablename__ = 'datapoints'
     id = db.Column(db.Integer, primary_key=True)
-    lat = db.Column(db.Float(precision=64), nullable=False)
-    lng = db.Column(db.Float(precision=64), nullable=False)
+    lat = db.Column(db.Float(precision=50), nullable=False)
+    lng = db.Column(db.Float(precision=50), nullable=False)
     type = db.Column(db.String, nullable=False)
     direction = db.Column(db.String, nullable=False)
